[test-core][Android] Move current implementation to legacy package
