+++ packages/expo-clipboard/android/src/test/java/expo/modules/clipboard/ClipboardModuleTest.kt
@@ -9,9 +9,9 @@ import android.os.Build
 import androidx.test.core.app.ApplicationProvider
 import expo.modules.kotlin.exception.CodedException
 import expo.modules.kotlin.exception.errorCodeOf
-import expo.modules.test.core.ModuleMock
-import expo.modules.test.core.ModuleMockHolder
-import expo.modules.test.core.assertCodedException
+import expo.modules.test.core.legacy.ModuleMock
+import expo.modules.test.core.legacy.ModuleMockHolder
+import expo.modules.test.core.legacy.assertCodedException
 import io.mockk.confirmVerified
 import io.mockk.verify
 import org.junit.Assert.assertEquals
