+++ packages/expo-crypto/android/src/test/java/expo/modules/crypto/CryptoModuleTest.kt
@@ -1,8 +1,8 @@
 package expo.modules.crypto
 
 import expo.modules.kotlin.exception.CodedException
-import expo.modules.test.core.ModuleMock
-import expo.modules.test.core.ModuleMockHolder
+import expo.modules.test.core.legacy.ModuleMock
+import expo.modules.test.core.legacy.ModuleMockHolder
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
