+++ packages/expo-media-library/android/src/test/java/expo/modules/medialibrary/albums/GetAlbumsTests.kt
@@ -10,9 +10,9 @@ import expo.modules.medialibrary.MockData
 import expo.modules.medialibrary.mockContentResolver
 import expo.modules.medialibrary.mockCursor
 import expo.modules.medialibrary.throwableContentResolver
-import expo.modules.test.core.PromiseMock
-import expo.modules.test.core.assertRejectedWithCode
-import expo.modules.test.core.promiseResolvedWithType
+import expo.modules.test.core.legacy.PromiseMock
+import expo.modules.test.core.legacy.assertRejectedWithCode
+import expo.modules.test.core.legacy.promiseResolvedWithType
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertThrows
 import org.junit.Before
