+++ packages/expo-media-library/android/src/test/java/expo/modules/medialibrary/assets/GetAssetInfoTests.kt
@@ -11,9 +11,9 @@ import expo.modules.medialibrary.MockData
 import expo.modules.medialibrary.mockContentResolver
 import expo.modules.medialibrary.mockContentResolverForResult
 import expo.modules.medialibrary.throwableContentResolver
-import expo.modules.test.core.PromiseMock
-import expo.modules.test.core.assertRejectedWithCode
-import expo.modules.test.core.promiseResolvedWithType
+import expo.modules.test.core.legacy.PromiseMock
+import expo.modules.test.core.legacy.assertRejectedWithCode
+import expo.modules.test.core.legacy.promiseResolvedWithType
 import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.just
